refactor: refactoring universal table to fulfill api requirements
