+++ src/config/routes.config.tsx
@@ -2,7 +2,7 @@ import { CompanyCreate } from '../components/pages/company/CompanyCreate';
 import { CompanyDetails } from '../components/pages/company/CompanyDetails';
 import { CompanyEdit } from '../components/pages/company/CompanyEdit';
 import { CompanyList } from '../components/pages/company/CompanyList';
-import { UserTable } from '../components/pages/company/UserTable';
+import { CompanyTable } from '../components/pages/company/CompanyTable';
 import { MainPage } from '../components/pages/MainPage';
 import { routes } from '../components/routes';
 
@@ -33,6 +33,6 @@ export const routeConfig = [
     },
     {
         path: routes.user.list(),
-        element: <UserTable />,
+        element: <CompanyTable />,
     },
 ];
